requirements.txt
openai>=1.0.0
instagrapi>=1.17.0
requests>=2.31.0
python-dotenv
Pillow==9.4.0
moviepy==1.0.3